refactored pug to load content for the user
