+++ app.js
@@ -20,13 +20,13 @@ const app = express();
 app.set("view engine", "pug");
 const store = new SequelizeStore({ db: sequelize });
 app.use(
-    session({
-        secret: sessionSecret,
-        name: "beavertown.sid",
-        store,
-        saveUninitialized: false,
-        resave: false,
-    })
+  session({
+    secret: sessionSecret,
+    name: "beavertown.sid",
+    store,
+    saveUninitialized: false,
+    resave: false,
+  })
 );
 
 app.use(logger("dev"));
@@ -41,25 +41,22 @@ app.use(restoreUser);
 // create Session table if it doesn't already exist
 store.sync();
 app.use(modalRouter);
-app.use(indexRouter);
-app.use("/users", usersRouter);
 app.use("/api", apiRouter);
 
-
 // catch 404 and forward to error handler
-app.use(function(req, res, next) {
-    next(createError(404));
+app.use(function (req, res, next) {
+  next(createError(404));
 });
 
 // error handler
-app.use(function(err, req, res, next) {
-    // set locals, only providing error in development
-    res.locals.message = err.message;
-    res.locals.error = req.app.get("env") === "development" ? err : {};
+app.use(function (err, req, res, next) {
+  // set locals, only providing error in development
+  res.locals.message = err.message;
+  res.locals.error = req.app.get("env") === "development" ? err : {};
 
-    // render the error page
-    res.status(err.status || 500);
-    res.render("error");
+  // render the error page
+  res.status(err.status || 500);
+  res.render("error");
 });
 
 module.exports = app;
