+++ public/stylesheets/modal.css
@@ -131,6 +131,9 @@ button {
   border: 1px red solid;
   width: 100%;
   height: 80%;
+}
+.project__main {
+  height: 741px;
   overflow-x: hidden;
   overflow-y: auto;
 }
@@ -145,6 +148,7 @@ button {
 }
 .project__header {
   display: flex;
+  position: sticky;
   justify-content: space-between;
   align-items: center;
 }
