+++ routes/modal.js
@@ -2,11 +2,8 @@ const express = require("express");
 const {
   csrfProtection,
   asyncHandler,
-  lowerCase,
-  upperCase,
-  oneNumeric,
-  alphaNumeric,
-  eightCharacters,
+  userValidators,
+  loginValidators,
 } = require("./utils.js");
 const db = require("../db/models");
 const { loginUser, logoutUser } = require("../auth");
@@ -19,98 +16,46 @@ router.get(
   "/",
   csrfProtection,
   asyncHandler(async (req, res, next) => {
+    let user = {};
+    let projects = {};
+    let tasks = { taskTitle: "Make a new Task" };
+    let projectDisplay = { projectName: "Project Title" };
+    let navBarTasks = { inProgress: 0, dueToday: 0, completed: 0 };
+
+    if (req.session.auth) {
+      const userId = req.session.auth.userId;
+      user = await db.User.findByPk(userId);
+      projects = await db.Project.findAll({
+        where: {
+          userId,
+        },
+      });
+      const projectId = projects[0].id;
+      projectDisplay = projects[0];
+      tasks = await db.Task.findAll({ where: { projectId } });
+    }
+
     res.render("modal", {
       title: "Test Modal",
-      csrfToken: req.csrfToken(),
-    });
-  })
-);
-
-router.get(
-  "/modal",
-  csrfProtection,
-  asyncHandler(async (req, res, next) => {
-    const user = await db.User.build();
-
-    res.render("modal", {
-      title: "Sign-Up",
       user,
+      projects,
+      projectDisplay,
+      tasks,
+      navBarTasks,
       csrfToken: req.csrfToken(),
     });
   })
 );
-
-const userValidators = [
-  check("firstName")
-    .exists({ checkFalsy: true })
-    .withMessage("Please provide a first name")
-    .isLength({ max: 50 })
-    .withMessage("First name cannot be more than 50 characters long"),
-  check("lastName")
-    .exists({ checkFalsy: true })
-    .withMessage("Please provide a last name")
-    .isLength({ max: 50 })
-    .withMessage("Last name cannot be more than 50 characters long"),
-  check("email")
-    .exists({ checkFalsy: true })
-    .withMessage("Please provide an email")
-    .isLength({ max: 255 })
-    .withMessage("email cannot be more than 255 characters long")
-    .custom((value) => {
-      return db.User.findOne({ where: { email: value } }).then((user) => {
-        if (user) {
-          return Promise.reject(
-            "The provided Email Address is already in use by another account"
-          );
-        }
-      });
-    }),
-  check("hashedPassword")
-    .exists({ checkFalsy: true })
-    .withMessage("Please Provide a password")
-    .isLength({ max: 255 })
-    .withMessage("password must be less than 255 characters")
-    .matches(lowerCase)
-    .withMessage(
-      "Please input a password with at least one lower case character"
-    )
-    .matches(upperCase)
-    .withMessage(
-      "Please input a password with at least one upper case character"
-    )
-    .matches(oneNumeric)
-    .withMessage("Please input a password with at least one numeric character")
-    .matches(alphaNumeric)
-    .withMessage(
-      "Please input a password with at least one alpha numeric character"
-    )
-    .matches(eightCharacters)
-    .withMessage("Please input a password at least eight characters long"),
-
-  check("confirmPassword")
-    .exists({ checkFalsy: true })
-    .withMessage("Please provide a value for Confirm Password")
-    .isLength({ max: 255 })
-    .withMessage("Confirm Password must not be more than 255 characters long")
-    .custom((value, { req }) => {
-      if (value !== req.body.hashedPassword) {
-        throw new Error("Confirm Password does not match Password");
-      }
-      return true;
-    }),
-  check("phoneNumber")
-    .isLength({ max: 10 })
-    .withMessage("Please input phone number with at most 10 digits"),
-  check("occupation")
-    .isLength({ max: 255 })
-    .withMessage("Max character limit is 255 characters"),
-];
 
 router.post(
   "/sign-up",
   csrfProtection,
   userValidators,
   asyncHandler(async (req, res, next) => {
+    let projects = {};
+    let tasks = { taskTitle: "Make a new Task" };
+    let projectDisplay = { projectName: "Project Title" };
+    let navBarTasks = { inProgress: 0, dueToday: 0, completed: 0 };
     const {
       firstName,
       lastName,
@@ -139,33 +84,33 @@ router.post(
       user.hashedPassword = hash;
       await user.save();
       loginUser(req, res, user);
-      res.redirect("/modal");
+      res.redirect("/");
     } else {
       const errorArray = errors.array().map((error) => error.msg);
       res.render("modal", {
         title: "Sign-Up",
         errorArray,
         user,
+        projects,
+        projectDisplay,
+        tasks,
+        navBarTasks,
         csrfToken: req.csrfToken(),
       });
     }
   })
 );
 
-const loginValidators = [
-  check("email")
-    .exists({ checkFalsy: true })
-    .withMessage("Please provide your login email address."),
-  check("hashedPassword")
-    .exists({ checkFalsy: true })
-    .withMessage("Don't forget your password!"),
-];
-
 router.post(
   "/login",
   csrfProtection,
   loginValidators,
   asyncHandler(async (req, res, next) => {
+    let user = {};
+    let projects = {};
+    let tasks = { taskTitle: "Make a new Task" };
+    let projectDisplay = { projectName: "Project Title" };
+    let navBarTasks = { inProgress: 0, dueToday: 0, completed: 0 };
     const { email, hashedPassword } = req.body;
     let errors = [];
     const loginErrors = validationResult(req);
@@ -180,7 +125,7 @@ router.post(
         );
         if (passwordMatch) {
           loginUser(req, res, user);
-          req.session.save(() => res.redirect("/modal"));
+          req.session.save(() => res.redirect("/"));
 
           return;
         }
@@ -193,15 +138,21 @@ router.post(
       title: "Home",
       email,
       errors,
+      user,
+      projects,
+      projectDisplay,
+      tasks,
+      navBarTasks,
       csrfToken: req.csrfToken(),
     });
   })
 );
+
 router.post(
   "/logout",
   asyncHandler(async (req, res) => {
     logoutUser(req, res);
-    req.session.save(() => res.redirect("/modal"));
+    req.session.save(() => res.redirect("/"));
   })
 );
