--- views/index.pug
@@ -1,22 +1,0 @@
-extends layout
-include utils.pug
-
-append head
-  //- add page specific styles by appending to the head
-  link(rel="stylesheet" href="/stylesheets/index.css")
-  //- add page specific js
-  script(src="/javascripts/index.js" type="module" defer)
-
-block content
-  h1= title
-  p Welcome to your Project Skeleton
-    +errorsSummary(errors)
-    form(action='/login' method='POST')
-      input(type='hidden' name='_csrf' value=csrfToken)
-      +form('Email Address','email','','email','email@email.com')
-      +form('Password', 'hashedPassword', '', 'password')
-      div
-       button(type='submit') Login
-  if locals.authenticated
-    form(action='/logout' method='post')
-      button(type='submit') Logout
