+++ views/layout.pug
@@ -2,7 +2,8 @@ doctype html
 html
   head
       block head
-        title="Express Starter App"
-        link(rel='stylesheet' href='/stylesheets/reset.css')
+        title="Task Beaver"
+        link(rel='stylesheet' href='/stylesheets/modal.css')
+        script(src="/javascripts/index.js" type="module")
   body
     block content
